launcher

+++ launcher.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 echo "welcome to kelboy-launcher"
 pid=$(ps -aux | grep -i pngview | awk '{print $2}')
-command2="kill -9 $pid"
-$command2
+if [$pid > 0]
+then
+  echo "killing old battery process"
+  command2="kill -9 $pid"
+  $command2
+fi
 
 cd bin
 ./battery.sh &
